pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ingestionator"
version = "0.1.0"
description = "Data ingestion pipeline"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "boto3",
    "duckdb",
    "pandas",
    "pyyaml",
    "pyarrow",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "setuptools",
    "ruff",
]

[tool.setuptools.packages.find]
include = ["ingestors", "ingestors.*", "sources", "sources.*", "targets", "targets.*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
